style(modeling) split sphere out from ellipsoid code
